pyproject.toml
[tool.poetry]
name = "parts-webapi"
version = "0.1.0"
description = "Putnam Area Robotics Team Web API"
authors = ["Brandon <brandon@bduke.dev>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
aiohappyeyeballs = "2.6.1"
aiohttp = "3.12.15"
aiosignal = "1.4.0"
asgiref = "3.9.1"
attrs = "25.3.0"
build = "1.3.0"
certifi = "2025.8.3"
cffi = "2.0.0"
charset-normalizer = "3.4.3"
click = "8.2.1"
cloudinary = "1.44.1"
cryptography = "45.0.7"
dj-database-url = "3.0.1"
django = "5.2.8"
django-cors-headers = "4.8.0"
django-filter = "25.1"
django-simple-history = "3.10.1"
django-webpush = "0.3.6"
djangorestframework = "3.16.1"
djangorestframework-simplejwt = "5.5.1"
frozenlist = "1.7.0"
gunicorn = "23.0.0"
http-ece = "1.2.1"
idna = "3.10"
multidict = "6.6.4"
packaging = "25.0"
pillow = "11.3.0"
pip-tools = "7.5.0"
py-vapid = "1.9.2"
pycparser = "2.23"
pyjwt = "2.10.1"
pyproject-hooks = "1.2.0"
pytz = "2025.2"
pywebpush = "2.0.3"
requests = "2.32.5"
six = "1.17.0"
sqlparse = "0.5.3"
typing-extensions = "4.15.0"
urllib3 = "2.5.0"
yarl = "1.20.1"
python-dotenv = "1.1.1"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pipdeptree = "^2.23.4"
pytest = "^8.0.0"
pytest-django = "^4.8.0"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
freezegun = "^1.5.0"
factory-boy = "^3.3.0"
requests-mock = "^1.12.0"
faker = "^33.0.0"


[tool.poetry.group.wvnet]
optional = true

[tool.poetry.group.wvnet.dependencies]
gunicorn = "23.0.0"
uwsgi = "2.0.30"
mysqlclient = "2.2.7"

[tool.poetry.group.uat]
optional = true

[tool.poetry.group.uat.dependencies]
gunicorn = "23.0.0"
uwsgi = "2.0.30"
psycopg2-binary = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
